requirements.txt
tensorflow
pandas
numpy
dvc
notebook
tqdm
Flask
Flask-cors
matplotlib
seaborn
python-box==6.0.2
pyYAML
ensure==1.0.2
joblib
scipy
types-PyYAML
-e .